- GpuEngine.cs: keep thin wrapper around DirectGpuTensorEngine
- DenseLayer.cs: merge using statements and add UpdateInputShape call

+++ src/Regression/DecisionTreeRegressionBase.cs
@@ -347,7 +347,8 @@ public abstract class DecisionTreeRegressionBase<T> : ITreeBasedRegression<T>
         // Serialize options
         writer.Write(Options.MaxDepth);
         writer.Write(Options.MinSamplesSplit);
-        writer.Write(double.IsNaN(Options.MaxFeatures) ? -1 : Options.MaxFeatures);
+        // Write MaxFeatures as int (-1 for NaN, cast to int otherwise)
+        writer.Write(double.IsNaN(Options.MaxFeatures) ? -1 : (int)Options.MaxFeatures);
         writer.Write(Options.Seed ?? -1);
         writer.Write((int)Options.SplitCriterion);
         // Serialize feature importances
